test: export mocks to dedicated package
